feat: rename CheckboxGroup to OptionList
